# frontend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: AMF API
  version: 0.1.0
  description: Application API for AMF.
paths:
  /api/auth/email-verification/confirm/:
    post:
      operationId: auth_email_verification_confirm_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Verified'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/email-verification/request/:
    post:
      operationId: auth_email_verification_request_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accepted'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/login/:
    post:
      operationId: auth_login_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/logout/:
    post:
      operationId: auth_logout_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      security:
      - cookieAuth: []
      - {}
      responses:
        '204':
          description: No response body
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/password-reset/confirm/:
    post:
      operationId: auth_password_reset_confirm_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '204':
          description: No response body
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/password-reset/request/:
    post:
      operationId: auth_password_reset_request_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetRequestRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accepted'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/register/:
    post:
      operationId: auth_register_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      security:
      - cookieAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/session/:
    get:
      operationId: auth_session_retrieve
      tags:
      - auth
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
          description: ''
  /api/health/:
    get:
      operationId: health_retrieve
      tags:
      - health
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
          description: ''
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnavailableHealth'
          description: ''
  /api/projects/:
    get:
      operationId: projects_list
      tags:
      - projects
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
    post:
      operationId: projects_create
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectWriteRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/projects/{id}/:
    get:
      operationId: projects_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - projects
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
    patch:
      operationId: projects_update
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProjectWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProjectWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProjectWriteRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/projects/{id}/activate/:
    post:
      operationId: projects_activate
      description: Activate the saved project.
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - projects
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/projects/{id}/deactivate/:
    post:
      operationId: projects_deactivate
      description: Return the saved project to draft.
      parameters:
      - in: header
        name: X-CSRFToken
        schema:
          type: string
        description: Current csrftoken cookie value for every unsafe request.
        required: true
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - projects
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
components:
  schemas:
    Accepted:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AcceptedStatusEnum'
      required:
      - status
    AcceptedStatusEnum:
      enum:
      - accepted
      type: string
      description: '* `accepted` - accepted'
    CodeEnum:
      enum:
      - validation_error
      - invalid_credentials
      - invalid_invitation
      - invalid_or_expired_token
      - not_authenticated
      - csrf_failed
      - email_verification_required
      - not_found
      - method_not_allowed
      - rate_limited
      type: string
      description: |-
        * `validation_error` - validation_error
        * `invalid_credentials` - invalid_credentials
        * `invalid_invitation` - invalid_invitation
        * `invalid_or_expired_token` - invalid_or_expired_token
        * `not_authenticated` - not_authenticated
        * `csrf_failed` - csrf_failed
        * `email_verification_required` - email_verification_required
        * `not_found` - not_found
        * `method_not_allowed` - method_not_allowed
        * `rate_limited` - rate_limited
    CurrencyEnum:
      enum:
      - KZT
      - USD
      - EUR
      type: string
      description: |-
        * `KZT` - Kzt
        * `USD` - Usd
        * `EUR` - Eur
    ErrorDetail:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/CodeEnum'
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/ErrorItem'
      required:
      - code
      - message
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
    ErrorItem:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
    Health:
      type: object
      properties:
        status:
          type: string
          readOnly: true
        database:
          type: string
          readOnly: true
      required:
      - database
      - status
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - password
    PasswordResetConfirmRequest:
      type: object
      properties:
        uid:
          type: string
          minLength: 1
        token:
          type: string
          minLength: 1
        new_password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - new_password
      - token
      - uid
    PasswordResetRequestRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
      required:
      - email
    PatchedProjectWriteRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        investment_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
          nullable: true
        currency:
          allOf:
          - $ref: '#/components/schemas/CurrencyEnum'
          default: KZT
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          readOnly: true
        description:
          type: string
          readOnly: true
        investment_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
        currency:
          allOf:
          - $ref: '#/components/schemas/CurrencyEnum'
          readOnly: true
        status:
          allOf:
          - $ref: '#/components/schemas/ProjectStatusEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        activated_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - activated_at
      - created_at
      - currency
      - description
      - id
      - investment_amount
      - status
      - title
      - updated_at
    ProjectListEnvelope:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
      required:
      - projects
    ProjectStatusEnum:
      enum:
      - draft
      - active
      type: string
      description: |-
        * `draft` - Draft
        * `active` - Active
    ProjectWriteRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        investment_amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,18}(?:\.\d{0,2})?$
          nullable: true
        currency:
          allOf:
          - $ref: '#/components/schemas/CurrencyEnum'
          default: KZT
      required:
      - title
    RegisterRequest:
      type: object
      properties:
        invitation_token:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - invitation_token
      - password
    Session:
      type: object
      properties:
        authenticated:
          type: boolean
        user:
          allOf:
          - $ref: '#/components/schemas/User'
          nullable: true
      required:
      - authenticated
      - user
    TokenRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    UnavailableHealth:
      type: object
      properties:
        status:
          type: string
          readOnly: true
        database:
          type: string
          readOnly: true
      required:
      - database
      - status
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        email_verified:
          type: boolean
          readOnly: true
      required:
      - email
      - email_verified
      - id
    UserEnvelope:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
      - user
    Verified:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/VerifiedStatusEnum'
      required:
      - status
    VerifiedStatusEnum:
      enum:
      - verified
      type: string
      description: '* `verified` - verified'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid